feat: reorganize the logger formatters, code optimizations
