+++ electron/common/logger/format/pretty.formatter.ts
@@ -4,40 +4,59 @@ import { Chalk } from 'chalk';
 import moment from 'moment';
 import type { LogData, LogFormatter, LogMessage } from '../types.js';
 import { LogLevel } from '../types.js';
-import { formatLogData } from './format-log-data.js';
-import { maskLogData } from './mask-log-data.js';
+import { formatLogData, maskLogData } from './format.utils.js';
 
-export const prettyLogFormatter = (options: {
+export const prettyLogFormatterFactory = (options: {
   colors: boolean;
 }): LogFormatter => {
   const { colors } = options;
 
   const chalk = new Chalk({ level: colors ? 1 : 0 });
 
-  const formatter: LogFormatter = (datas: Array<LogMessage>): string => {
-    const logLines = datas.map((data) => {
-      const formattedData = maskLogData(formatLogData(data)) as LogMessage;
-
-      const { timestamp, level, scope, message, ...rest } = formattedData;
-
-      const timeStr = formatTime({ chalk, timestamp });
-      const levelStr = formatLevel({ chalk, level });
-      const scopeStr = formatScope({ chalk, scope });
-      const messageStr = formatMessage({ chalk, message });
-      const restStr = formatObject({ data: rest, colors });
-
-      // Because of how the colorized text is created or rendered,
-      // whitespaces behave like in HTML: multiple spaces are collapsed to one.
-      // To preserve whitespace, you must add the extra spaces to the string
-      // that is passed to one of the chalk methods.
-
-      return `${timeStr} ${levelStr} ${scopeStr} ${messageStr} ‣ ${restStr}`;
-    });
+  const formatter: LogFormatter = (logMessages: Array<LogMessage>): string => {
+    const logLines = formatLogMessages({ chalk, logMessages });
     return logLines.join('\n') + '\n';
   };
 
   return formatter;
 };
+
+const formatLogMessages = (options: {
+  chalk: ChalkInstance;
+  logMessages: Array<LogMessage>;
+}): Array<string> => {
+  const { chalk, logMessages } = options;
+
+  return logMessages.map((logMessage) => {
+    return formatLogMessage({ chalk, logMessage });
+  });
+};
+
+const formatLogMessage = (options: {
+  chalk: ChalkInstance;
+  logMessage: LogMessage;
+}): string => {
+  const { chalk, logMessage } = options;
+
+  // Leaky abstraction to know that we can cast to LogMessage from LogData...
+  // But doing so makes it much simpler to pull out the fields we need next.
+  const formattedData = maskLogData(formatLogData(logMessage)) as LogMessage;
+
+  const { timestamp, level, scope, message, ...rest } = formattedData;
+
+  const timeStr = formatTime({ chalk, timestamp });
+  const levelStr = formatLevel({ chalk, level });
+  const scopeStr = formatScope({ chalk, scope });
+  const messageStr = formatMessage({ chalk, message });
+  const restStr = formatObject({ data: rest, colors: chalk.level > 0 });
+
+  return `${timeStr} ${levelStr} ${scopeStr} ${messageStr} ‣ ${restStr}`;
+};
+
+// Because of how the colorized text is created or rendered,
+// whitespaces behave like in HTML: multiple spaces are collapsed to one.
+// To preserve whitespace, you must add the extra spaces to the string
+// that is passed to one of the chalk methods.
 
 const formatTime = (options: {
   chalk: ChalkInstance;
