+++ electron/common/sanitize/__tests__/sanitize-utils.test.ts
@@ -1,43 +1,7 @@
 import { describe, expect, it } from 'vitest';
-import {
-  isNotMaskable,
-  maskLogData,
-  maskSensitiveValues,
-} from '../mask-log-data.js';
+import { maskSensitiveValues } from '../sanitize.utils.js';
 
-describe('mask-log-data', () => {
-  describe('#maskLogData', () => {
-    const data: Record<string, any> = {
-      accessToken: 'accessToken1',
-      password: 'password1',
-      apiKey: 'apiKey1',
-      credential: 'credential1',
-      nested: {
-        accessToken: 'accessToken2',
-        password: 'password2',
-        apiKey: 'apiKey2',
-        credential: 'credential2',
-      },
-    };
-
-    it('masks password, accessToken, and apiKey properties by default', () => {
-      const result = maskLogData(data);
-
-      expect(result).toEqual({
-        accessToken: '***REDACTED***',
-        password: '***REDACTED***',
-        apiKey: '***REDACTED***',
-        credential: 'credential1',
-        nested: {
-          accessToken: '***REDACTED***',
-          password: '***REDACTED***',
-          apiKey: '***REDACTED***',
-          credential: 'credential2',
-        },
-      });
-    });
-  });
-
+describe('sanitize-utils', () => {
   describe('#maskSensitiveValues', () => {
     const data: Record<string, any> = {
       accessToken: 'accessToken1',
@@ -111,35 +75,25 @@ describe('mask-log-data', () => {
         },
       });
     });
-  });
 
-  describe('#isNotMaskable', () => {
-    it('returns true for null', () => {
-      expect(isNotMaskable(null)).toBe(true);
-    });
-
-    it('returns true for undefined', () => {
-      expect(isNotMaskable(undefined)).toBe(true);
-    });
-
-    it('returns true for Date', () => {
-      expect(isNotMaskable(new Date())).toBe(true);
-    });
-
-    it('returns true for string', () => {
-      expect(isNotMaskable('string')).toBe(true);
-    });
-
-    it('returns true for number', () => {
-      expect(isNotMaskable(42)).toBe(true);
-    });
-
-    it('returns true for boolean', () => {
-      expect(isNotMaskable(true)).toBe(true);
-    });
-
-    it('returns false for object', () => {
-      expect(isNotMaskable({ foo: 42 })).toBe(false);
+    it.each([
+      undefined,
+      null,
+      new Date(),
+      'string',
+      42,
+      true,
+      false,
+      [],
+      [1],
+      new Map(),
+      new Set(),
+    ])('does not mask non-objects: %s', (nonObject: any) => {
+      expect(
+        maskSensitiveValues({
+          object: nonObject,
+        })
+      ).toBe(nonObject);
     });
   });
 });
