+++ electron/common/sanitize/__tests__/tsconfig.json
@@ -1,0 +1,3 @@
+{
+  "extends": "../../tsconfig.test.json"
+}
